requirements.txt
streamlit
TTS
Pillow
opencv-python
ffmpeg-python